requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
cssutils==2.11.1
edgartools==5.22.0
filelock==3.25.0
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpxthrottlecache==0.3.0
humanize==4.15.0
idna==3.11
Jinja2==3.1.6
loguru==0.7.3
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
nest-asyncio==1.6.0
numpy==2.4.3
orjson==3.11.7
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyrate-limiter==4.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
rank-bm25==0.2.2
RapidFuzz==3.14.3
requests==2.32.5
rich==14.3.3
sec-downloader==0.11.2
sec-edgar-downloader==5.1.0
sec-parser==0.58.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
stamina==25.2.0
tabulate==0.9.0
tenacity==9.1.4
textdistance==4.6.3
tqdm==4.67.3
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
xxhash==3.6.0
